backend-services/requirements.txt
# StEx 后端服务 - Python 依赖
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
httpx>=0.27.0

# 数据采集
akshare>=1.14.0
tushare>=1.2.89

# AI / 工作流（Moonshot API + 互联网搜索）
openai>=1.50.0
duckduckgo-search>=6.0.0
feedparser>=6.0.0
# crewai>=0.80.0
# langgraph>=0.2.0
